llm_provider_standalone/requirements.txt
# LLMProvider Dependencies
litellm>=1.0.0
pydantic>=2.0.0
jinja2>=3.0.0
python-dotenv>=1.0.0
